Started A binomial probability calculator
Created a binomial probability calculator. The main experiment with this program was to attempt a rudimentary command tag system similar to how windows powershell scripting works. Parsing wasn't a hugely difficult issue but I'm not entirely sure I used the correct formula for what I technically want this program to do. Regardless, the program supports two different types of data entry by the tags -ip and -it (inputting percentage and time respectively). There is a delimiting character read in user input (~) which allows for large clumps of data to be entered at any one given time and parsed. Future ideas to progress this program are to take text files as an input and also further clean up the istreamstring to vector array methods.